moved generic category menu component to routes home component
