+++ crwn-clothing/package.json
@@ -9,6 +9,7 @@
     "css-loader": "^6.7.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "6",
     "react-scripts": "5.0.1",
     "sass": "^1.57.1",
     "sass-loader": "^13.2.0",
