+++ crwn-clothing/src/index.scss
@@ -1,6 +1,6 @@
 body {
   margin: 0;
-  font-family: "Open Sans Condensed"
+  font-family: "Open Sans Condensed",
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
